up :)  On branch next  Changes to be committed: 	modified:   cmd/gomem/commands.go 	modified:   gomem.go

+++ cmd/gomem/commands.go
@@ -54,7 +54,8 @@ func confirm(msg string) bool {
 	return false
 }
 
-/// commands
+/// commands ///
+// status //
 func la() (string, error) {
 	var str string
 	for key, v := range igs.Gmap {
@@ -71,6 +72,38 @@ func ls() (string, error) {
 	}
 	return str, nil
 }
+func state() (string, error) {
+	var str string
+	str += color.GreenString(fmt.Sprintln("igs.dir:", igs.GetDir()))
+	infos, err := ioutil.ReadDir(igs.GetDir())
+	if err == nil {
+		for _, info := range infos {
+			if info.IsDir() {
+				str += color.HiGreenString(fmt.Sprintln("sub category:", info.Name()))
+			}
+		}
+	}
+	for key, v := range igs.Gmap {
+		str += color.GreenString("%s:", key)
+		str += fmt.Sprint("[", v.JSON.Title, "]")
+		str += fmt.Sprintln(":read only", !v.Override)
+	}
+	return str, nil
+}
+func show(s string) (string, error) {
+	if !strings.HasSuffix(s, ".json") {
+		s = s + ".json"
+	}
+	g, ok := igs.Gmap[s]
+	if !ok {
+		return "not found:" + s, nil
+	}
+	str := color.CyanString(fmt.Sprintln("[", g.JSON.Title, "]"))
+	str += color.CyanString(fmt.Sprintln(g.JSON.Content))
+	return str, nil
+}
+
+// contact to data //
 func newGomem() (string, error) {
 	fpath := read("filename:>")
 	g, err := gomem.New(fpath, true)
@@ -96,36 +129,12 @@ func newGomemWithName(s string) (string, error) {
 	}
 	return "new gomem included", nil
 }
-func write() (string, error) {
-	b := confirm(fmt.Sprintf("write mems into %s", igs.GetDir()))
-	var result string
-	if b {
-		for key, x := range igs.Gmap {
-			if err := x.WriteFile(); err != nil {
-				result += fmt.Sprintln(key, err.Error())
-			}
-		}
-		return result, nil
+func include() (string, error) {
+	err := igs.IncludeJSON()
+	if err != nil {
+		return err.Error(), nil
 	}
-	return "", nil
-}
-func state() (string, error) {
-	var str string
-	str += color.GreenString(fmt.Sprintln("igs.dir:", igs.GetDir()))
-	infos, err := ioutil.ReadDir(igs.GetDir())
-	if err == nil {
-		for _, info := range infos {
-			if info.IsDir() {
-				str += color.HiGreenString(fmt.Sprintln("sub category:", info.Name()))
-			}
-		}
-	}
-	for key, v := range igs.Gmap {
-		str += color.GreenString("%s:", key)
-		str += fmt.Sprint("[", v.JSON.Title, "]")
-		str += fmt.Sprintln(":read only", !v.Override)
-	}
-	return str, nil
+	return "data cache reincluded: from " + igs.GetDir(), nil
 }
 func cd() (string, error) {
 	if !confirm("cd is dropped all data cache") {
@@ -145,7 +154,7 @@ func cd() (string, error) {
 	igs = tmpgs
 	return "changed directory to:" + igs.GetDir(), nil
 }
-func show(s string) (string, error) {
+func modContent(s string) (string, error) {
 	if !strings.HasSuffix(s, ".json") {
 		s = s + ".json"
 	}
@@ -153,9 +162,57 @@ func show(s string) (string, error) {
 	if !ok {
 		return "not found:" + s, nil
 	}
-	str := color.CyanString(fmt.Sprintln("[", g.JSON.Title, "]"))
-	str += color.CyanString(fmt.Sprintln(g.JSON.Content))
+	msg := color.GreenString("%s:[%s]:", s, g.JSON.Title) + color.MagentaString("%s\n", g.JSON.Content)
+	c := read(msg + "mod content:>")
+	g.JSON.Content = c
+	return color.GreenString("content modified"), nil
+}
+func removeCache(s string) (string, error) {
+	if !strings.HasSuffix(s, ".json") {
+		s = s + ".json"
+	}
+	if _, ok := igs.Gmap[s]; !ok {
+		return "not found:" + s, nil
+	}
+	if confirm("remove cache:"+s) == false {
+		return "", nil
+	}
+	delete(igs.Gmap, s)
+	return color.RedString("removed cache:" + s), nil
+}
+func todo() (string, error) {
+	var str string
+	for key, g := range igs.Gmap {
+		if strings.HasPrefix(key, "todo"+string(filepath.Separator)) {
+			str += color.GreenString("%s:", key)
+			str += color.CyanString("[%s]\n", g.JSON.Title)
+			str += color.MagentaString("%s\n", g.JSON.Content)
+		}
+	}
 	return str, nil
+}
+
+// physical //
+func makeSubcategory(s string) (string, error) {
+	subname := filepath.Join(igs.GetDir(), filepath.Base(s))
+	err := os.Mkdir(subname, 0777)
+	if err != nil {
+		return err.Error(), nil
+	}
+	return "maked subcategory:" + subname, nil
+}
+func write() (string, error) {
+	b := confirm(fmt.Sprintf("write mems into %s", igs.GetDir()))
+	var result string
+	if b {
+		for key, x := range igs.Gmap {
+			if err := x.WriteFile(); err != nil {
+				result += fmt.Sprintln(key, err.Error())
+			}
+		}
+		return result, nil
+	}
+	return "", nil
 }
 func remove(s string) (string, error) {
 	if !strings.HasSuffix(s, ".json") {
@@ -165,20 +222,15 @@ func remove(s string) (string, error) {
 	if err != nil {
 		return err.Error(), nil
 	}
+	if confirm("remove:"+fullpath) == false {
+		return "", nil
+	}
 	err = os.Remove(fullpath)
 	if err != nil {
 		return err.Error(), nil
 	}
 	delete(igs.Gmap, s)
 	return color.RedString(fullpath + " is removed"), nil
-}
-func makeSubcategory(s string) (string, error) {
-	subname := filepath.Join(igs.GetDir(), filepath.Base(s))
-	err := os.Mkdir(subname, 0777)
-	if err != nil {
-		return err.Error(), nil
-	}
-	return "maked subcategory:" + subname, nil
 }
 func removeSubcategory(s string) (string, error) {
 	subname := filepath.Join(igs.GetDir(), filepath.Base(s))
@@ -209,19 +261,23 @@ func interactive(r io.Reader, w io.Writer, prefix string, gs *gomem.Gomems) erro
 	interWriter = w
 
 	sub := gomem.SubNewWithBase(r, w)
-	sub.Addf("la", la, "show Gmap")
-	sub.Addf("ls", ls, "ls Gmap keys")
+	sub.Addf("la", la, "show gs.Gmap")
+	sub.Addf("ls", ls, "ls gs.Gmap keys")
+	sub.Addf(":q", gomem.Exit, "exit alias")
+	sub.Addf("state", state, "show state of gs")
 	sub.Addf("new", newGomem, "new gomem")
 	sub.Addf("write", write, "write all data to gs.dir")
-	sub.Addf("state", state, "show state of gs")
-	sub.Addf("cd", cd, "change working directory")
-	sub.Addf(":q", gomem.Exit, "exit alias")
+	sub.Addf("cd", cd, "change working directory, and exchange of data cache")
+	sub.Addf("todo", todo, "show subcategory [todo/*]")
+	sub.Addf("include", include, "reinclude from gs.dir")
 
-	sub.Addfa("mkdir", makeSubcategory, "mkdir make subcategory in igs.dir")
 	sub.Addfa("show", show, "show title and content")
-	sub.Addfa("remove", remove, "remove json cache")
-	sub.Addfa("removesub", removeSubcategory, "remove subcategory")
+	sub.Addfa("mkdir", makeSubcategory, "mkdir make subcategory in gs.dir")
+	sub.Addfa("remove", remove, "remove json file")
+	sub.Addfa("removesub", removeSubcategory, "remove subcategory directory")
+	sub.Addfa("removecache", removeCache, "remove cache data")
 	sub.Addfa("new", newGomemWithName, "")
+	sub.Addfa("modc", modContent, "modify content")
 
 	if err := sub.Repl(prefix); err != nil {
 		return err
